more test fixes

+++ rsocket-core/src/test/java/io/rsocket/frame/FrameHeaderFlyweightTest.java
@@ -65,7 +65,7 @@ public class FrameHeaderFlyweightTest {
     assertEquals(
         4,
         FrameHeaderFlyweight.decodeMetadataLength(
-            byteBuf, FrameHeaderFlyweight.FRAME_HEADER_LENGTH));
+            byteBuf, FrameHeaderFlyweight.FRAME_HEADER_LENGTH).longValue());
   }
 
   @Test
